backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
httpx==0.24.1
python-dotenv==1.0.0
PyJWT==2.8.0
pydantic==1.10.12
python-multipart==0.0.6
asyncpg
